+++ releasenotes.md
@@ -15,6 +15,7 @@ Release Notes
   the `LocalTime` property on `MaxMind.MinFraud.Response.Device`.
 * Added support for the `/credit_card/is_virtual` output. This is exposed as
   the `IsVirtual` property on `MaxMind.MinFraud.Response.CreditCard`.
+* Added `PayoutChange` to the `EventType` enum.
 
 1.6.0 (2018-01-22)
 ------------------