+++ releasenotes.md
@@ -11,6 +11,11 @@ Release Notes
   * `Dalenys`
   * `Oney`
   * `Posconnect`
+* Added support for the `/device/local_time` output. This is exposed as
+  the `LocalTime` property on `MaxMind.MinFraud.Response.Device`.
+* Added support for the `/credit_card/is_virtual` output. This is exposed as
+  the `IsVirtual` property on `MaxMind.MinFraud.Response.CreditCard`.
+* Added `PayoutChange` to the `EventType` enum.
 
 1.6.0 (2018-01-22)
 ------------------